Explicitly cast res_call to fix test failure

+++ testsuite/libffi.closures/cls_multi_sshort.c
@@ -65,7 +65,7 @@ int main (void)
   /* { dg-output "2 32765: 32767" } */
   printf("res: %d\n", (unsigned short)res_call);
   /* { dg-output "\nres: 32767" } */
-  CHECK(res_call == 32767);
+  CHECK((unsigned short)res_call == 32767);
 
   CHECK(ffi_prep_closure_loc(pcl, &cif, test_func_gn, NULL, code)  == FFI_OK);
